Lots of compiler warnings removed.
git-svn-id: https://svn.aros.org/svn/aros/trunk@54880 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/dopus/Config/screen_colors.c
@@ -44,7 +44,7 @@ int pen,prop;
         GetRGB32(vp->ColorMap,pen,1,cbuf);
         for (a=0;a<3;a++) {
             scalecol(&cbuf[a]);
-            lsprintf(palette_buf[a],"%ld",cbuf[a]);
+            lsprintf(palette_buf[a],"%ld",(long int)cbuf[a]);
             if (prop) FixSliderPot(Window,&palettegads[a],cbuf[a],1<<bpg,1,1);
             ShowSlider(Window,&palettegads[a]);
             RefreshGList(&palettegads[a+3],Window,NULL,1);
